# nextlinux_engine/services/apiext/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "This is the Anchore Engine API. Provides the primary external API for users of the service."
  version: "0.1.20"
  title: "Anchore Engine API Server"
  contact:
    email: "nurmi@nextlinux.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
#externalDocs:
#  description: "Anchore Engine"
#  url: "http://github.com/nextlinux/nextlinux-engine"
...
basePath: "/"
schemes:
  - "http"
  - "https"
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /system/feeds/{feed}:
    put:
      tags:
        - System
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.system
      operationId: toggle_feed_enabled
      x-nextlinux-authz-action: updateFeeds
      description: "Disable the feed so that it does not sync on subsequent sync operations"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: feed
          in: path
          required: true
          type: string
        - name: enabled
          type: boolean
          required: true
          in: query
      responses:
        200:
          description: "FeedInfo"
          schema:
            $ref: "#/definitions/FeedMetadata"
        500:
          description: "Internal server error processing the request. Retry expected"
        400:
          description: "Bad request, fix and resend"
    delete:
      tags:
        - System
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.system
      operationId: delete_feed
      x-nextlinux-authz-action: updateFeeds
      description: "Delete the groups and data for the feed and disable the feed itself"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: feed
          in: path
          required: true
          type: string
      responses:
        200:
          description: "Successfully deleted"
        500:
          description: "Internal server error processing the request. Retry expected"
        404:
          description: "Not found"
  /system/feeds/{feed}/{group}:
    put:
      tags:
        - System
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.system
      x-nextlinux-authz-action: updateFeeds
      operationId: toggle_group_enabled
      description: "Disable a specific group within a feed to not sync"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: feed
          in: path
          required: true
          type: string
        - name: group
          in: path
          type: string
          required: true
        - name: enabled
          type: boolean
          required: true
          in: query
      responses:
        200:
          description: "FeedInfo listing"
          schema:
            type: array
            items:
              $ref: "#/definitions/FeedMetadata"
        500:
          description: "Internal server error processing the request. Retry expected"
        400:
          description: "Bad request, fix and resend"
    delete:
      tags:
        - System
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.system
      operationId: delete_feed_group
      x-nextlinux-authz-action: updateFeeds
      description: "Delete the group data and disable the group itself"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: feed
          in: path
          required: true
          type: string
        - name: group
          in: path
          type: string
          required: true
      responses:
        200:
          description: "Successfully deleted"
        500:
          description: "Internal server error processing the request. Retry expected"
        404:
          description: "Not found"
...
definitions:
...
  FeedMetadata:
    type: object
    description: Metadata on a single feed based on what the engine finds from querying the endpoints
    properties:
      name:
        type: string
        description: name of the feed
      created_at:
        type: string
        format: date-time
        description: Date the metadata record was created in engine (first seen on source)
      updated_at:
        type: string
        format: date-time
        description: Date the metadata was last updated
      groups:
        type: array
        items:
          $ref: "#/definitions/FeedGroupMetadata"
      last_full_sync:
        type: string
        format: date-time
  FeedGroupMetadata:
    type: object
    properties:
      name:
        type: string
      created_at:
        type: string
        format: date-time
      last_sync:
        type: string
        format: date-time
      record_count:
        type: integer
...